requirements.txt
requests
bs4
argparse
tqdm
tenacity
pyrogram
tgcrypto
python-dotenv
flask
